refactor: move all logging to components into monitoring dir
